+++ GateKeeper.Server/Database/MySqlConnectorWrapper.cs
@@ -9,7 +9,7 @@ namespace GateKeeper.Server.Database
 
 
     // Concrete class implementing the interface
-    public class MySqlConnectorWrapper : IMySqlConnectorWrapper, IAsyncDisposable
+    public class MySqlConnectorWrapper : IMySqlConnectorWrapper
     {
         private readonly MySqlConnection _connection;
 
@@ -80,57 +80,4 @@ namespace GateKeeper.Server.Database
         }
 
     }
-
-    public class MySqlDataReaderWrapper : IMySqlDataReaderWrapper, IAsyncDisposable
-    {
-        private readonly MySqlDataReader _reader;
-
-        public MySqlDataReaderWrapper(MySqlDataReader reader)
-        {
-            _reader = reader;
-        }
-
-        public async Task<bool> ReadAsync(CancellationToken cancellationToken = default)
-        {
-            return await _reader.ReadAsync(cancellationToken);
-        }
-
-        public object this[string name] => _reader[name];
-
-        public int GetInt32(string name)
-        {
-            return _reader.GetInt32(_reader.GetOrdinal(name));
-        }
-
-        public string GetString(string name)
-        {
-            return _reader.GetString(_reader.GetOrdinal(name));
-        }
-
-        public DateTime GetDateTime(string name)
-        {
-            return _reader.GetDateTime(_reader.GetOrdinal(name));
-        }
-
-        public bool IsDBNull(int ordinal)
-        {
-            return _reader.IsDBNull(ordinal);
-        }
-
-        public int GetOrdinal(string name)
-        {
-            return _reader.GetOrdinal(name);
-        }
-
-        public async Task<bool> NextResultAsync(CancellationToken cancellationToken = default)
-        {
-            return await _reader.NextResultAsync(cancellationToken);
-        }
-
-        public ValueTask DisposeAsync()
-        {
-            return _reader.DisposeAsync();
-        }
-    }
-
 }
