+++ GateKeeper.Server/Interface/IMySqlConnectorWrapper.cs
@@ -13,16 +13,4 @@ namespace GateKeeper.Server.Interface
         void CloseConnection();
         Task<Dictionary<string, object>> ExecuteNonQueryWithOutputAsync(string commandText, CommandType commandType, params MySqlParameter[] parameters);
     }
-
-    public interface IMySqlDataReaderWrapper : IAsyncDisposable
-    {
-        Task<bool> ReadAsync(CancellationToken cancellationToken = default);
-        object this[string name] { get; }
-        int GetInt32(string name);
-        string GetString(string name);
-        DateTime GetDateTime(string name);
-        bool IsDBNull(int ordinal);
-        int GetOrdinal(string name);
-        Task<bool> NextResultAsync(CancellationToken cancellationToken = default);
-    }
 }
